hrapp-on-confidential-cloud/02-acl-ledger-python/requirements.txt
azure.identity==1.4.0
pygtail==0.11.1
